FabricCommon(): Move more shared code into this class.
Moving the following shared code into FabricCommon():
- FabricReplacedCommon(). _config_deploy()
- FabricUpdateCommon(). _config_deploy()

- FabricReplacedCommon(). _config_save()
- FabricUpdateCommon(). _config_save()

Also:

FabricCreateCommon(): Minor update to a docstring.
